test(PaymentController): PayController Integrate Test
