+++ src/test/java/sejong/coffee/yun/repository/pay/PayJpaRepositoryTest.java
@@ -47,7 +47,7 @@ public class PayJpaRepositoryTest extends BeforeCreatedData {
         //given
         String paymentKey = "5zJ4xY7m0kODnyRpQWGrN2xqGlNvLrKwv1M9ENjbeoPaZdL6";
 
-        CardPayment approvalPayment = CardPayment.approvalPayment(cardPayment, paymentKey, LocalDateTime.now());
+        CardPayment approvalPayment = CardPayment.approvalPayment(cardPayment, paymentKey, LocalDateTime.now().toString());
 
         //when
         CardPayment save = jpaPayRepository.save(approvalPayment);
@@ -61,12 +61,12 @@ public class PayJpaRepositoryTest extends BeforeCreatedData {
         //given
         String paymentKey = "5zJ4xY7m0kODnyRpQWGrN2xqGlNvLrKwv1M9ENjbeoPaZdL6";
 
-        CardPayment approvalPayment = CardPayment.approvalPayment(cardPayment, paymentKey, LocalDateTime.now());
+        CardPayment approvalPayment = CardPayment.approvalPayment(cardPayment, paymentKey, LocalDateTime.now().toString());
 
         jpaPayRepository.save(approvalPayment);
 
         //when
-        Optional<CardPayment> result = jpaPayRepository.findByOrderIdAnAndPaymentStatus(uuid, PaymentStatus.DONE);
+        Optional<CardPayment> result = jpaPayRepository.findByOrderUuidAnAndPaymentStatus(uuid, PaymentStatus.DONE);
 
         //then
         assertThat(result.isPresent()).isTrue();
@@ -77,7 +77,7 @@ public class PayJpaRepositoryTest extends BeforeCreatedData {
         //given
         String paymentKey = "5zJ4xY7m0kODnyRpQWGrN2xqGlNvLrKwv1M9ENjbeoPaZdL6";
 
-        CardPayment approvalPayment = CardPayment.approvalPayment(cardPayment, paymentKey, LocalDateTime.now());
+        CardPayment approvalPayment = CardPayment.approvalPayment(cardPayment, paymentKey, LocalDateTime.now().toString());
 
         jpaPayRepository.save(approvalPayment);
 
@@ -95,9 +95,9 @@ public class PayJpaRepositoryTest extends BeforeCreatedData {
         String paymentKey_2 = "5zJ4xY7m0kODnyRpQWGrN2xqGlNvLrKwv1M9ENjbeoPaZdL7";
         String paymentKey_3 = "5zJ4xY7m0kODnyRpQWGrN2xqGlNvLrKwv1M9ENjbeoPaZdL8";
 
-        CardPayment approvalPayment_1 = CardPayment.approvalPayment(cardPayment, paymentKey_1, LocalDateTime.now());
-        CardPayment approvalPayment_2 = CardPayment.approvalPayment(cardPayment, paymentKey_2, LocalDateTime.now());
-        CardPayment approvalPayment_3 = CardPayment.approvalPayment(cardPayment, paymentKey_3, LocalDateTime.now());
+        CardPayment approvalPayment_1 = CardPayment.approvalPayment(cardPayment, paymentKey_1, LocalDateTime.now().toString());
+        CardPayment approvalPayment_2 = CardPayment.approvalPayment(cardPayment, paymentKey_2, LocalDateTime.now().toString());
+        CardPayment approvalPayment_3 = CardPayment.approvalPayment(cardPayment, paymentKey_3, LocalDateTime.now().toString());
 
         jpaPayRepository.save(approvalPayment_1);
         jpaPayRepository.save(approvalPayment_2);
